Refactor calendar into modular components
